pyproject.toml
[tool.poetry]
name = "test-task-django-rocketdata"
version = "0.1.0"
description = ""
authors = ["Dmitry <kiselevdv1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
django = ">=3.2,<4.0"
djangorestframework = ">=3.12,<4.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
celery = ">=5.2,<6.0"
django-redis = ">=5.4.0,<6.0.0"
django-celery-beat = ">=2.7.0,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
gunicorn = ">=23.0.0,<24.0.0"
whitenoise = ">=6.9.0,<7.0.0"
qrcode = "^8.1"
pillow = "^11.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"